+++ trunk/src/userprog/process.h
@@ -13,7 +13,7 @@ void process_activate (void);
 
 void start_process (void *file_name_);
 bool setup_stack (void **esp);
-bool load_segment (struct file *file, int ofs, uint32_t *upage,  
+bool load_segment (struct file *file, int ofs, uint8_t *upage,  
                           uint32_t read_bytes, uint32_t zero_bytes,
                           bool writable);
 bool install_page (void *upage, void *kpage, bool writable);
